See the result of the game
I fixed up the innerText logic in the code, I realized that the element wasn't being printed so in each if, else if statement I just added a createElement('p'). To get the result printed, I created a div and used a querySelector to get printed result in that div.